requirements.txt
python-dotenv
requests
pint
sympy
scipy
numpy
plotly
streamlit
pydantic
